execution/bot/requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
gspread==6.1.2
google-auth==2.34.0
google-api-python-client==2.149.0
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
